changed policy of scraping the most recent debates
Only the debate parts older than 12 hours are scraped and existing
speeches are not re-created when scraping debate parts of an
existing sitting.

The point is not to re-create (and change id) of any existing
entity. There is a drawback of this approach - if an debate part
older than 12 hours is modified on source web, the changes will
not be scraped.